+++ just_audio_web/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.4.11
 
 * Bump package:web upper bound to <0.6.0
-* Add AudioPlayer.setWebCrossOrigin for CORS on web.
+* Add AudioPlayer.setWebCrossOrigin for CORS on web (@danielwinkler).
 
 ## 0.4.10